systemd: remove obsolete bbappend
The default value of PACKAGECONFIG in upstream systemd recipe already
contain networkd and resolved.

This bbappend use += to add those values, and the original recipe uses
a weak default value (??=), which means that those weak defaults will
not be used. Instead PACKAGECONFIG will only contain networkd and
resolved which is not the intention.

Signed-off-by: Erik Botö <erik.boto@pelagicore.com>